Refactor methods out of logger into win_io module
Add tests
